Draw value of items on trade screen

+++ tactics_project/src/com/marklynch/objects/GameObjectTemplate.java
@@ -34,7 +34,7 @@ public class GameObjectTemplate implements InventoryParent {
 	public boolean canBePickedUp = true;
 	public boolean decorative = false;
 
-	public float value = 1;
+	public int value = 1;
 	public Date pickUpdateDateTime = new Date();
 	public float widthRatio = 1;
 	public float heightRatio = 1;
